+++ abci/did/v1/ndid.go
@@ -33,30 +33,30 @@ import (
 )
 
 var isNDIDMethod = map[string]bool{
-	"InitNDID":                              true,
-	"RegisterNode":                          true,
-	"AddNodeToken":                          true,
-	"ReduceNodeToken":                       true,
-	"SetNodeToken":                          true,
-	"SetPriceFunc":                          true,
-	"AddNamespace":                          true,
-	"DisableNamespace":                      true,
-	"SetValidator":                          true,
-	"AddService":                            true,
-	"DisableService":                        true,
-	"UpdateNodeByNDID":                      true,
-	"UpdateService":                         true,
-	"RegisterServiceDestinationByNDID":      true,
-	"DisableNode":                           true,
-	"DisableServiceDestinationByNDID":       true,
-	"EnableNode":                            true,
-	"EnableServiceDestinationByNDID":        true,
-	"EnableNamespace":                       true,
-	"EnableService":                         true,
-	"SetTimeOutBlockRegisterMsqDestination": true,
-	"AddNodeToProxyNode":                    true,
-	"UpdateNodeProxyNode":                   true,
-	"RemoveNodeFromProxyNode":               true,
+	"InitNDID":                         true,
+	"RegisterNode":                     true,
+	"AddNodeToken":                     true,
+	"ReduceNodeToken":                  true,
+	"SetNodeToken":                     true,
+	"SetPriceFunc":                     true,
+	"AddNamespace":                     true,
+	"DisableNamespace":                 true,
+	"SetValidator":                     true,
+	"AddService":                       true,
+	"DisableService":                   true,
+	"UpdateNodeByNDID":                 true,
+	"UpdateService":                    true,
+	"RegisterServiceDestinationByNDID": true,
+	"DisableNode":                      true,
+	"DisableServiceDestinationByNDID":  true,
+	"EnableNode":                       true,
+	"EnableServiceDestinationByNDID":   true,
+	"EnableNamespace":                  true,
+	"EnableService":                    true,
+	"SetTimeOutBlockRegisterIdentity":  true,
+	"AddNodeToProxyNode":               true,
+	"UpdateNodeProxyNode":              true,
+	"RemoveNodeFromProxyNode":          true,
 }
 
 func initNDID(param string, app *DIDApplication, nodeID string) types.ResponseDeliverTx {
@@ -767,16 +767,16 @@ func enableService(param string, app *DIDApplication, nodeID string) types.Respo
 	return ReturnDeliverTxLog(code.OK, "success", "")
 }
 
-func setTimeOutBlockRegisterMsqDestination(param string, app *DIDApplication, nodeID string) types.ResponseDeliverTx {
-	app.logger.Infof("SetTimeOutBlockRegisterMsqDestination, Parameter: %s", param)
-	var funcParam TimeOutBlockRegisterMsqDestination
+func setTimeOutBlockRegisterIdentity(param string, app *DIDApplication, nodeID string) types.ResponseDeliverTx {
+	app.logger.Infof("SetTimeOutBlockRegisterIdentity, Parameter: %s", param)
+	var funcParam TimeOutBlockRegisterIdentity
 	err := json.Unmarshal([]byte(param), &funcParam)
 	if err != nil {
 		return ReturnDeliverTxLog(code.UnmarshalError, err.Error(), "")
 	}
 
-	key := "TimeOutBlockRegisterMsqDestination"
-	var timeOut data.TimeOutBlockRegisterMsqDestination
+	key := "TimeOutBlockRegisterIdentity"
+	var timeOut data.TimeOutBlockRegisterIdentity
 	timeOut.TimeOutBlock = funcParam.TimeOutBlock
 	// Check time out block > 0
 	if timeOut.TimeOutBlock <= 0 {